+++ src/components/swerve_wheel.py
@@ -10,7 +10,7 @@ from wpimath.geometry import Rotation2d
 from wpimath.kinematics import SwerveModulePosition, SwerveModuleState
 
 from magicbot import will_reset_to
-from lemonlib.preference import SmartPreference, SmartProfile
+from lemonlib.smart.preference import SmartPreference, SmartProfile
 
 
 class SwerveWheel:
